+++ models/contacts.json
@@ -94,5 +94,11 @@
     "name": "Ala ma teraz nosorożca",
     "email": "wufwuf@dog.com",
     "phone": "777 777 777"
+  },
+  {
+    "id": "1712094530256",
+    "name": "Marcel",
+    "email": "olaboga@coto.com",
+    "phone": "(23)23-2323-23"
   }
 ]
